[#237] refactor: PostSearchController를 생성하여 검색 관련 API 이동
